requirements.txt
pyinstaller
python-dotenv
openai
mistralai
google-generativeai
anthropic